chore: get rid of unused functions
